update minor problems
