+++ practice-app/group5_api/exchangeRateAPI/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class ExchangeRateAPIConfig(AppConfig):
+    name = 'exchangeRateAPI'
