+++ practice-app/group5_api/group5_api/settings.py
@@ -37,7 +37,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'myapp'
+    'exchangeRateAPI'
 ]
 
 MIDDLEWARE = [
