+++ practice-app/group5_api/home/templates/home/home.html
@@ -1,0 +1,23 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <center><h1>Group5 home page!</h1></center>
+    </head>
+    <body>
+        <center>
+            <form action="weatherCondition/" method="get">
+                <button>Search a weather condition</button>
+            </form>
+            <br>
+            <form action="dailyQuote" method="GET">
+                <button>Bring me sports quote of the day</button>
+            </form>
+            <br>
+            <form action="nbaStats/" method="get">
+                <button>Search NBA Players</button>
+            </form>
+            <br>
+        </center>
+    </body>
+</html>
+
