+++ practice-app/group5_api/home/urls.py
@@ -1,0 +1,8 @@
+from django.urls import path, include
+
+from . import views
+
+urlpatterns = [
+    path('', views.index, name='index'),
+    path('exchangeRateAPI/', include('exchangeRateAPI.urls'))
+]
