+++ practice-app/group5_api/home/views.py
@@ -1,0 +1,8 @@
+from django.shortcuts import render
+from rest_framework.decorators import api_view
+
+
+@api_view(['GET'])
+def index(request):
+    return render(request,'home/home.html',{})
+
